# MetricsManager/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Metric collection agent service API",
    "description": "Provides collection data\r\n                        Loading processor, Free Memory\r\n                        Using hard disk, Return speed over the network\r\n                        Heap size  from registered agents",
    "termsOfService": "https://example.com/terms",
    "contact": {
      "name": "Orlikov Fedor",
      "url": "https://orlfi.tk",
      "email": "orlfi@mail.ru"
    },
    "license": {
      "name": "MIT",
      "url": "https://example.com/license"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Agents/register": {
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Registers a new agent",
        "description": "Sample request:\r\n            \r\n    GET api/agents/register",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAgentCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If everything is ok"
          }
        }
      }
    },
    "/api/Agents/enable/{agentId}": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Enables agent",
        "description": "Sample request:\r\n            \r\n    GET api/enable/1",
        "parameters": [
          {
            "name": "AgentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If everything is ok"
          }
        }
      }
    },
    "/api/Agents/disable/{agentId}": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Disables agent",
        "description": "Sample request:\r\n            \r\n    GET api/disable/1",
        "parameters": [
          {
            "name": "AgentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If everything is ok"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterAgentCommand": {
        "type": "object",
        "properties": {
          "AgentUrl": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "IsEnabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}